Make DMM use instances for core functions, add global instance wrappers
