+++ src/main_test.h
@@ -1,0 +1,11 @@
+#ifndef DMM_MAIN_TEST_H
+#define DMM_MAIN_TEST_H
+
+#include <stddef.h>
+#include <dmm.h>
+#include "main.h"
+
+size_t test_dmm_malloc();
+size_t test_dmm_free_sets_header();
+
+#endif
